Autoconfigure
Next, this autoconfigure key is a brand new feature. When a service is autoconfigured, it means that Symfony will automatically tag it when possible. For example, our MarkdownExtension extends \Twig_Extension:

 35 lines  src/AppBundle/Twig/MarkdownExtension.php
... lines 1 - 6
class MarkdownExtension extends \Twig_Extension {
... lines 9 - 33
}
Which implements Twig_ExtensionInterface:

abstract class Twig_Extension implements Twig_ExtensionInterface {
    // ...
}
That's actually the important part. When a service is autoconfigured and its class implements Twig_ExtensionInterface, the twig.extension tag is automatically added for you.
Basically, Symfony is saying:

Hey! I see you configured a service that implements Twig_ExtensionInterface. Obviously, that's a Twig extension, so let me configure it for you.

This works for many - but not all tags. It does not work for doctrine.event_subscriber or form.type_extension:

 41 lines  app/config/services.yml
... lines 1 - 5
services:
... lines 7 - 21
  app.doctrine.hash_password_listener:
    class: AppBundle\Doctrine\HashPasswordListener
    tags:
      - { name: doctrine.event_subscriber }
  app.form.help_form_extenion:
    class: AppBundle\Form\TypeExtension\HelpFormExtension
    tags:
      - { name: form.type_extension, extended_type: Symfony\Component\Form\Extension\Core\Type\FormType }
... lines 31 - 41
Because it has an extended_type tag option... which the system can't guess for you. When you're developing a feature, the docs will tell you whether or not you need to add the tag manually. If you do add a tag, even though you didn't need to, no problem! Your tag takes precedence.

So, all our services are autowired and autoconfigured! But, it doesn't make any difference, besides shortening our config just a little. And when we refresh, everything still works!