Updating all of the Service
Let's repeat that for the rest of our services. Honestly, when we upgraded KnpUniversity, this was the most tedious step: going one-by-one, copying each service id, copying the class name, and then adding it to legacy_aliases.yml. We wrote a dirty script that used the Yaml class to load services.yml and at least create the legacy_aliases.yml file for us.

Notice that LoginFormAuthenticator has no configuration anymore! Cool! Just set it to a ~

Multiple Services for the Same Class

 39 lines  app/config/services.yml
... lines 1 - 8
services:
... lines 10 - 29
  app.encouraging_message_generator:
    class: AppBundle\Service\MessageGenerator
... lines 32 - 34
  app.discouraging_message_generator:
     class: AppBundle\Service\MessageGenerator
... lines 37 - 39
The last two services are a problem: we can't set the id to the class name, because the class is the same for each! When this happens, use the old id naming convention. We're going to talk more about this situation soon.

And, we're done! We just changed all of the service ids to class names... but thanks to legacy_aliases.yml, our code won't break. This may not have felt significant, but it was actually a big step forward. Now, we can talk about private services.