Controller Action Injection To fix our error, we need to use classic dependency injection. And actually, there are two ways to do this in a controller. First, we could of course, go to the top of our class, add a __construct function, type-hint a MarkdownTransformer argument set that on a property, and use it below. Thanks to autowiring, we wouldn't need to touch any config files.
But, because this is a bit tedious and so common to do in a controller, we've added a shortcut. In controllers only, you can autowire a service into an argument of your action method. We'll add MarkdownTransformer {dollar_sign}markdownTransformer.

Now, remove the {dollar_sign}this->get() line... which is a shortcut for {dollar_sign}this->container->get().

This fixes things... because we've eliminated the {dollar_sign}container->get() call that does not work with private services.

Celebrate by removing the first alias! The rest are easy!