The app.markdown_extension id isn't referenced anywhere, so remove that. The app.security.login_form_authenticator is used in two places: security.yml and also UserController.
Copy the new service id - the class name. In security.yml, just replace the old with the new.
Next, in UserController, I'll search for authenticator. Ah, we're fetching it out of the container directly! We know the fix: type-hint a new argument with LoginFormAuthenticator {dollar_sign}authenticator and use that below