Our last public service is MessageManager. Now we know how to fix this. In GenusAdminController, find editAction() and add an argument: MessageManager {dollar_sign}messageManager Use that below in both places. Back in services.yml, make that service private.
This is reason to celebrate! All our services are now private! We're using proper dependency injection on everything! Thanks to this, the container will optimize itself for performance and give us clear errors if we make a mistake... anywhere.

Next, we need to talk more about aliases: the key to unlocking the full potential of autowiring.