Let's talk more about what happens when autowiring goes wrong. Right now, the MarkdownTransformer class has two arguments, type-hinted with MarkdownParserInterface and Cache from Doctrine:
 35 lines  src/AppBundle/Service/MarkdownTransformer.php
... lines 1 - 4
use Doctrine\Common\Cache\Cache;
use Knp\Bundle\MarkdownBundle\MarkdownParserInterface;
class MarkdownTransformer {
... lines 10 - 12
  public function __construct(MarkdownParserInterface {dollar_sign}markdownParser, Cache {dollar_sign}cache){
... lines 15 - 16
  }
... lines 18 - 33
}
But, these are not being autowired: we're explicitly specifying each argument.

Letting Symfony tell you about Autowiring Failures

If I were creating this service today, I actually wouldn't specify any configuration at first. Nope, I'd just create the class, add the type-hints, and let Symfony tell me if it had any problems autowiring my arguments.

To show off one of the ways autowiring can fail, I'm going to remove the first argument and set it to an empty string.

When you do this, it means that you do want Symfony to try to autowire it. Symfony will try to autowire the first argument, but not the second. Actually, if this looks ugly to you, I agree! In a few minutes, I'll show you a cleaner way of explicitly wiring only some arguments.

Anyways, let's see if the MarkdownParserInterface type-hint can be autowired. Refresh the page... or any page.

Explosion!

Cannot autowire service AppBundle\Service\MarkdownTransformer: argument .
It's very clearly saying that there is a problem with this specific argument.

The error continues:

... it references interface MarkdownParserInterface but no such service exists.
This is because it's looking for a service with exactly this id. But, none was found! So, it tries to help out:

You should maybe alias this interface to one of these existing services.
and it lists one, two, three, four, five services in the container that implement MarkdownParserInterface. This is autowiring the Symfony way: there's no guess work.