Using Aliases to add Valid Autowiring Types
Previously, we were using a service called markdown.parser. Find your terminal and get some information about this service:

php bin/console debug:container markdown.parser

Interesting! This is actually an alias to markdown.parser.max. This service comes from KnpMarkdownBundle, and it ships with a few different markdown parsers. It then creates an alias from markdown.parser to whatever parser we configured as the default.

So to fix the error, we have two options. First, we could of course explicitly specify the argument, just like we were doing before. Or, as the error suggests, we can create an alias. Let's do that: alias.

We just told Symfony exactly what service to autowire when it sees the MarkdownParserInterface type-hint. In theory, KnpMarkdownBundle would come with this alias already, and it probably will in the future.