Solution 1) Fixing the Type-Hint
Here's the first question: is there a different type-hint that we should be using instead for this service? Let's find out!

Head to your terminal. We already know about the debug:container command:

php bin/console debug:container
This gives us a big list of every public service in the container. The blue text is the id of each service. But guess what? Service id's are much less important in Symfony 3.3... because we almost always rely on type-hints and autowiring.

Re-run the command again with a new --types option:

php bin/console debug:container --types
Voilà! This is a list of all valid type-hints that you can use for autowiring. This is awesome. If you search for encoder, you'll find one called UserPasswordEncoderInterface. This is the type-hint we should use! Symfony ships with this alias to enable autowiring.

Cool! Let's find out where we're using this:

git grep UserPasswordEncoder
Two places: HashPasswordListener and LoginFormAuthenticator. Open up HashPasswordListener. Then, add Interface to the end of the use statement, and also the type-hint.

That's it.

Open up LoginFormAuthenticator and do the exact same thing: update the use... and the argument.

Ok, go back to the browser! Refresh, and watch those 10 deprecations. Bam! 8 deprecations!

If you check the list now, we still have one more autowiring deprecation. This time, apparently, it's unhappy about an EntityManager type-hint.

Same question as before: is there a better type-hint to use? Let's find out:

php bin/console debug:container --types

Search for EntityManager and... boom! There is an EntityManagerInterface alias. This is the officially supported type-hint.