We saw earlier that sometimes you only need to pass one argument... and the rest can be autowired. For MarkdownTransformer, we only need to configure the second argument:
 45 lines  app/config/services.yml
... lines 1 - 5
services:
... lines 7 - 30
  AppBundle\Service\MarkdownTransformer:
    arguments: ['', '@doctrine_cache.providers.my_markdown_cache']
... lines 33 - 45
To allow the first to be continue to be autowired, we set it to an empty string. That works... but it's just weird. So, in Symfony 3.3, there's a better way.

In MarkdownTransformer, the argument names are {dollar_sign}markdownParser and {dollar_sign}cache:

 35 lines  src/AppBundle/Service/MarkdownTransformer.php
... lines 1 - 7
class MarkdownTransformer {
... lines 10 - 12
  public function __construct(MarkdownParserInterface {dollar_sign}markdownParser, Cache {dollar_sign}cache){
... lines 15 - 16
  }
... lines 18 - 33
}
The  argument is the one we need to configure. Back in services.yml, copy the cache service id, clear out the arguments, and add {dollar_sign}cache: '@doctrine_cache.providers.my_markdown_cache'

The dollar sign is the important part: it tells Symfony that we're actually configuring an argument by its name. {dollar_sign}cache here must match {dollar_sign}cache here.

And it works beautifully - refresh now! Everything is happy!