Named Arguments are Validated But wait! Isn't this dangerous!? I mean, normally, if I were coding in MarkdownTransformer, I could safely rename this variable to {dollarSign}cacheDriver
That change shouldn't make any difference to any code outside of this class.

But suddenly now... it does make a difference! The  in services.yml no longer matches the argument name. Isn't this a huge problem!? Actually, no. Yes, you will get an error - you can see it when you refresh. But this error will happen if you try to refresh any page across your entire system. Symfony validates all of your services and configuration. And as soon as it saw {dollarSign}cache in services.yml with no corresponding argument, it screamed.