We've done a lot of work, and I showed you the ugliest parts of the new system so that you can solve them in your project. That's cool... but so far, coding hasn't been much fun!
And that's a shame! Once you're done upgrading, using the new configuration system is a blast. Let's take it for a legit test drive.

Creating an Event Subscriber

Here's the goal: create an event listener that adds a header to every response. Step 1: create an EventSubscriber directory - though this could live anywhere - and a file inside called AddNiceHeaderEventSubscriber.

Event subscribers always look the same: they must implement EventSubscriberInterface. I'll go to the Code-Generate menu, or Command+N on a Mac - and select 'Implement Methods' to add the one required function: public static getSubscribedEvents(). To listen to the kernel.response event, return KernelEvents::RESPONSE set to onKernelResponse. On top, create that method: onKernelResponse() with a FilterResponseEvent object: that's the argument passed to listeners of this event. Inside, add a header.

Ok, we've touched only one file and written 23 lines of code. And... we're done! Yep, this will already work. I'll open up my network tools, then refresh one more time. For the top request, click Headers, scroll down and... there it is! We just added an event subscriber to Symfony... by just creating the event subscriber. Yea... it kinda makes sense.

The new class was automatically registered as a service and automatically tagged thanks to autoconfigure.