So yes, changing the argument in your service class will break your app. But the error is unignorable: nothing works until you fix it. Update the configuration to {dollarSign}cacheDriver
And refresh again.

Back working! This is what makes Symfony's autowiring special. If there is any question or problem wiring the arguments to any service, Symfony throws an exception at compile team... meaning, it throws an exception when you try to refresh any page. This means no surprises: it's not possible to have an autowiring error on only one page and not notice it.